Scripts to MySQL import and merge
Do not remove the placeholder file index.html in uploads folder.

+++ admin/mysqlmerge.php
@@ -49,7 +49,9 @@ chmod("uploads", 0755);
 $scanfiles = scandir('uploads');
 foreach ($scanfiles as $num => $scanfile){
 	if ($scanfile !="." && $scanfile !=".."){
-		unlink("uploads/".$scanfile);
+        if( !strstr($scanfile, ".html") ) {
+		    unlink("uploads/".$scanfile);
+        }
 	}
 }
